// frontend/src/app/api/model/customer-response-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CustomerResponseDto { 
    id?: number;
    discountRate?: number;
    firstName?: string;
    lastName?: string;
    phone?: string;
}