// frontend/src/app/api/model/customer-create-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CustomerCreateDto { 
    discountRate: number;
    firstName: string;
    lastName: string;
    phone: string;
}