// frontend/src/app/api/model/customer-update-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CustomerUpdateDto { 
    discountRate?: number;
    firstName?: string;
    lastName?: string;
    phone?: string;
}